Make mk example
it replaces mkdir so it can make more stuff
